requirement.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
conda-pack==0.4.0
connected-components-3d==1.7.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.1
Deprecated==1.2.4
dicom2nifti==0.0.0
entrypoints==0.3
future==0.18.2
idna==2.9
imageio==2.8.0
joblib==0.14.1
keyring==12.0.2
keyrings.alt==3.1
kiwisolver==1.1.0
linecache2==1.0.0
matplotlib==3.2.2
MedPy==0.4.0
networkx==2.4
nibabel==3.0.1
numpy==1.16.0
opencv-python==4.1.1.26
pandas==1.0.1
Pillow==7.0.0
pycparser==2.20
pydicom==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
SecretStorage==2.3.1
SimpleITK==1.2.4
six==1.14.0
sklearn==0.0
synapseclient==2.0.0
threadpoolctl==2.0.0
torch==1.3.1+cu100
torchsummary==1.5.1
tqdm==4.42.1
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.25.8
wrapt==1.12.1
